GestureMgr::DrawSkeletonKinectData: the skeleton loop counter is signed
The loop-back test in build/373307D9/asm/system/gesture/GestureMgr.s is
`cmpwi cr6, r11, 0x6` (0x82DBB014), a SIGNED compare; we had written
`(unsigned int)i < NUM_SKELETONS`, which lowers to `cmplwi`. The cast was
ours, not the image's.

The replace row is gone (312 now equal) but the headline moves 94.00 ->
93.86: removing it re-shuffles one scheduling window further down
(`clrrwi`/`addi` ordering at 0x82DBB040). Taking the faithful spelling
deliberately; the residual on this function is FPR assignment
(target f28=0.0f/f30=1.0f/f29=RAD2DEG vs ours f29/f31/f28) and stack-slot
colouring around the shoulder Vector3 temporaries, neither source-reachable
from what I could see.

Also carries the InstanceGeomOwnerBones negative result: neither
`mGeomOwner.Ptr()` nor `mGeomOwner.Ptr() != nullptr` turns that function's
one remaining `cmpwi` into the image's `cmplwi` at 0x82642F6C.

+++ src/system/gesture/GestureMgr.cpp
@@ -535,7 +535,9 @@ void GestureMgr::DrawSkeletonKinectData() {
     Vector2 textPos(0.15f, 0.2f);
 
     auto& debugDir = mDebugDir;
-    for (int i = 0; (unsigned int)i < NUM_SKELETONS; i++) {
+    // Signed: the image's loop-back test is `cmpwi cr6, r11, 0x6` at 0x82DBB014,
+    // not `cmplwi` -- the (unsigned int) cast was ours.
+    for (int i = 0; i < NUM_SKELETONS; i++) {
         const Skeleton &skel = GetSkeleton(i);
         SkeletonTrackingState tracking = skel.TrackingState();
